Extract <Result> and <LoadingIndicator> from <SolFeesApp>
