+++ src/constants.js
@@ -1,0 +1,7 @@
+const SOL_PER_LAMPORT = 0.000000001
+const TX_CAP = 20000
+
+export {
+  SOL_PER_LAMPORT,
+  TX_CAP,
+}
